Moving a bunch of stuff.
