+++ analysis_and_sanity_checks/atom_types_only_experiments/analysis/create_trajectory_cif_files.py
