+++ analysis_and_sanity_checks/atom_types_only_experiments/analysis/plot_atom_type_probabilities.py
