+++ analysis_and_sanity_checks/atom_types_only_experiments/patches/fixed_starting_point_trajectory_initializer.py
