+++ analysis_and_sanity_checks/atom_types_only_experiments/training/run_diffusion.sh
