+++ analysis_and_sanity_checks/components_analysis/plot_target_conditional_score_different_sigmas.py
@@ -7,11 +7,12 @@ sigma, showing the behavior for different normalizations.
 import matplotlib.pyplot as plt
 import torch
 
+from analysis_and_sanity_checks.components_analysis import \
+    PLOTS_OUTPUT_DIRECTORY
 from diffusion_for_multi_scale_molecular_dynamics.analysis import (
     PLEASANT_FIG_SIZE, PLOT_STYLE_PATH)
 from diffusion_for_multi_scale_molecular_dynamics.score.wrapped_gaussian_score import \
     get_coordinates_sigma_normalized_score
-from experiments.components_analysis import PLOTS_OUTPUT_DIRECTORY
 
 plt.style.use(PLOT_STYLE_PATH)
