+++ analysis_and_sanity_checks/toy_problems/training/base_config.yaml
