+++ analysis_and_sanity_checks/toy_problems/training/consistency_regularizer/run.sh
