+++ analysis_and_sanity_checks/toy_problems/training/consistency_with_analytical_guide_regularizer/run.sh
