+++ analysis_and_sanity_checks/toy_problems/training/fokker_planck_regularizer/run.sh
