+++ analysis_and_sanity_checks/toy_problems/training/no_regularizer/run.sh
